chore(webgl): Split vertex-array, vao, renderbuffer into webgl and legacy classes
