+++ modules/webgl/src/adapter/objects/webgl-resource.ts
@@ -2,9 +2,9 @@
 import {Resource, assert, uid, stubRemovedMethods} from '@luma.gl/api';
 import type {Device} from '@luma.gl/api';
 export type {ResourceProps} from '@luma.gl/api';
-import {isWebGL2, assertWebGLContext} from '../context/context/webgl-checks';
-import {getKey, getKeyValue} from '../webgl-utils/constants-to-keys';
-import WebGLDevice from '../adapter/webgl-device';
+import {isWebGL2, assertWebGLContext} from '../../context/context/webgl-checks';
+import {getKey, getKeyValue} from '../../webgl-utils/constants-to-keys';
+import WebGLDevice from '../webgl-device';
 
 const ERR_RESOURCE_METHOD_UNDEFINED = 'Resource subclass must define virtual methods';
