+++ modules/webgl/src/classes/renderbuffer.ts
@@ -1,29 +1,9 @@
 /* eslint-disable no-inline-comments */
-import {assert} from '@luma.gl/api';
 import GL from '@luma.gl/constants';
 import WebGLDevice from '../adapter/webgl-device';
-import WebGLResource, {ResourceProps} from './webgl-resource';
-import {
-  isRenderbufferFormatSupported, getRenderbufferFormatBytesPerPixel
-} from '../adapter/converters/renderbuffer-formats';
-import {isWebGL2} from '../context/context/webgl-checks';
+import WEBGLRenderbuffer, {RenderbufferProps} from '../adapter/objects/webgl-renderbuffer';
 
-export type RenderbufferProps = ResourceProps & {
-  format: number;
-  width?: number;
-  height?: number;
-  samples?: number;
-};
-
-const DEFAULT_RENDERBUFFER_PROPS: Required<RenderbufferProps> = {
-  id: undefined,
-  handle: undefined,
-  userData: undefined,
-  format: 0,
-  width: 1, 
-  height: 1, 
-  samples: 0
-};
+export {RenderbufferProps}
 
 /**
  * Renderbuffers are GPU objects that contain images.
@@ -35,80 +15,9 @@ const DEFAULT_RENDERBUFFER_PROPS: Required<RenderbufferProps> = {
  * use Textures instead.
  * Renderbuffer objects also natively accommodate Multisampling (MSAA).
  */
-export class ImmutableRenderbuffer extends WebGLResource<RenderbufferProps> {
-  get [Symbol.toStringTag](): string { return 'Renderbuffer'; }
-
-  get width(): number { return this.props.width; }
-  get height(): number { return this.props.height; }
-  get format(): number  { return this.props.format; }
-  get samples(): number { return this.props.samples; }
-
+export default class Renderbuffer extends WEBGLRenderbuffer {
   static isSupported(gl: WebGLRenderingContext, options?: {format?: number}): boolean {
-    return !options.format || isRenderbufferFormatSupported(gl, options.format);
-  }
-
-  constructor(gl: WebGLRenderingContext, props: RenderbufferProps) {
-    super(WebGLDevice.attach(gl), props, DEFAULT_RENDERBUFFER_PROPS);
-    this._initialize(this.props);
-  }
-
-  // PRIVATE METHODS
-
-  /** Creates and initializes a renderbuffer object's data store */
-  protected _initialize(props: Required<RenderbufferProps>) {
-    const {format, width, height, samples} = props;
-    assert(format, 'Needs format');
-
-    this.trackDeallocatedMemory();
-
-    this.gl.bindRenderbuffer(GL.RENDERBUFFER, this.handle);
-
-    if (samples !== 0 && isWebGL2(this.gl)) {
-      // @ts-expect-error
-      this.gl.renderbufferStorageMultisample(GL.RENDERBUFFER, samples, format, width, height);
-    } else {
-      this.gl.renderbufferStorage(GL.RENDERBUFFER, format, width, height);
-    }
-
-    this.gl.bindRenderbuffer(GL.RENDERBUFFER, null);
-
-    this.trackAllocatedMemory(
-      width * height * (samples || 1) * getRenderbufferFormatBytesPerPixel(format)
-    );
-
-    return this;
-  }
-
-  // RESOURECE IMPLEMENTATION
-
-  _createHandle() {
-    return this.gl.createRenderbuffer();
-  }
-
-  _deleteHandle(): void {
-    this.gl.deleteRenderbuffer(this.handle);
-    this.trackDeallocatedMemory();
-  }
-
-  _bindHandle(handle): void {
-    this.gl.bindRenderbuffer(GL.RENDERBUFFER, handle);
-  }
-}
-
-
-/**
- * Renderbuffers are GPU objects that contain images.
- * In contrast to Textures they are optimized for use as render targets, with Framebuffers.
- * while Textures may not be, and are the logical choice when
- * you do not need to sample (i.e. in a post-pass shader)
- * from the produced image. If you need to resample
- * (such as when reading depth back in a second shader pass),
- * use Textures instead.
- * Renderbuffer objects also natively accommodate Multisampling (MSAA).
- */
-export default class Renderbuffer extends ImmutableRenderbuffer {
-  static isSupported(gl: WebGLRenderingContext, options?: {format?: number}): boolean {
-    return ImmutableRenderbuffer.isSupported(gl, options);
+    return WEBGLRenderbuffer.isSupported(gl, options);
   }
 
   static getSamplesForFormat(gl: WebGL2RenderingContext, options: {format: number}): number{
