+++ modules/webgl/src/classes/transform-feedback.ts
@@ -2,7 +2,7 @@ import {log, isObjectEmpty} from '@luma.gl/api';
 import GL from '@luma.gl/constants';
 import WebGLDevice from '../adapter/webgl-device';
 import {isWebGL2, assertWebGL2Context} from '../context/context/webgl-checks';
-import WebGLResource, {ResourceProps} from './webgl-resource';
+import WebGLResource, {ResourceProps} from '../adapter/objects/webgl-resource';
 import Buffer from './webgl-buffer';
 
 export type TransformFeedbackProps = ResourceProps & {
